requirements.txt
absl-py==0.11.0
astunparse==1.6.3
axial-positional-embedding==0.2.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.1.1
clang==5.0
colorama==0.4.4
contrastive-learner==0.1.0
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
dominate==2.6.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
joblib==0.17.0
jsonpatch==1.32
jsonpointer==2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.4.1
libclang==14.0.6
linear-attention-transformer==0.15.0
linformer==0.2.0
local-attention==1.1.1
Markdown==3.3.3
MarkupSafe==2.1.1
matplotlib==3.3.3
multiprocess==0.70.11.1
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
Pillow==8.0.1
product-key-memory==0.1.10
protobuf==3.13.0
psutil==5.9.2
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.2.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.6.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.11.18
torch==1.12.0+cu102
torchfile==0.1.0
torchvision==0.13.0+cu102
tornado==6.1
tqdm==4.53.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vector-quantize-pytorch==0.1.0
visdom==0.1.8.9
websocket-client==1.2.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlwt==1.3.0
zipp==3.4.0
